Move to a utils directory for toggles utils
