+++ src/app/containers/Toggle/utils/isToggleEnabled.js
